chore: move users and small tweaks
